Take a pass at creating a CommandHandler object to field Slack commands.
The custom /bgg route is set to handle the request data submitted by Slack. The new
CommandHandler object checks for the presence of a subcommand (e.g., 'latest'), and
attempts to process the request, assuming the other requirements of the commands are
met.

This change is still a work-in-progress because it's missing a lot of important
details, such as the need to verify that the source of the request has permission
to make the request at all, and there are some Python conventions I'm not yet familiar
with, such as my IDE warning me that the get_latest method in the CommandHandler class
could be static. In PHP, I basically avoid static methods wherever possible. In Python,
it appears that it's a common convention to mix classes and functions inside module
files, and as such, it may be necessary to move that method outside of the CommandHandler.

I'm not crazy about that change, but hey, it's the end of the day and I've reached a logical
stopping point and gotta commit this now. :)